Revised update and delete employee methods

+++ src/main/java/com/springproject/cruddemo/rest/EmployeeRestController.java
@@ -55,8 +55,16 @@ public class EmployeeRestController {
 
     // add mapping for DELETE /employees/{employeeId} - delete an employee
     @DeleteMapping("/employees/{employeeId}")
-    public void deleteEmployee(@PathVariable int employeeId) {
+    public String deleteEmployee(@PathVariable int employeeId) {
+        Employee tempEmployee = this.employeeService.findById(employeeId);
+
+        // throw exception if null
+        if (tempEmployee == null) {
+            throw new RuntimeException("Employee id not found - " + employeeId);
+        }
+
         this.employeeService.deleteById(employeeId);
+        return "Deleted employee id - " + employeeId;
     }
 
 }
